tests/Element-setAttribute.html

<!DOCTYPE html>
<title>Element.setAttribute</title>
<script src="testharness.js"></script>
<script src="attributes.js"></script>
<script src="productions.js"></script>
<div id="log"></div>
<script>
// XXX some of this is testing Attr and getAttribute / getAttributeNS
// maybe it should all be merged?
// Step 1
test(function() {
  var el = document.createElement("foo")
  for (var i = 0; i < invalid_names.length; i++) {
    assert_throws("INVALID_CHARACTER_ERR", function() { el.setAttribute(invalid_names[i], "test") })
  }
}, "When qualifiedName does not match the Name production, an " +
   "INVALID_CHARACTER_ERR exception is to be thrown.")

// Step 2
test(function() {
  var el = document.createElement("div")
  el.setAttribute("ALIGN", "left")
  assert_equals(el.getAttributeNS("", "ALIGN"), "")
  assert_equals(el.getAttributeNS("", "align"), "left")
  assert_equals(el.getAttribute("align"), "left")
  el.setAttribute("CHEEseCaKe", "tasty")
  assert_equals(el.getAttributeNS("", "CHEEseCaKe"), "")
  assert_equals(el.getAttributeNS("", "cheesecake"), "tasty")
  assert_equals(el.getAttribute("cheesecake"), "tasty")
}, "setAttribute should lowercase its name argument")

// Step 3
test(function() {
  var el = document.createElement("foo")
  var tests = ["xmlns", "xmlns:a", "xmlnsx", "xmlns0"]
  for (var i = 0; i < tests.length; i++) {
    assert_throws("NAMESPACE_ERR",
                  function() { el.setAttribute(tests[i], "fail") },
                  "Expected exception for " + tests[i] + ".")
  }
}, "")

// Step 4
test(function() {
  var el = document.createElement("foo")
  for (var i = 0; i < valid_names.length; i++) {
    el.setAttribute(valid_names[i], "test")
    assert_equals(el.getAttribute(valid_names[i]), "test")
  }
}, "Basic functionality should be intact.")

// Step 5
test(function() {
  var el = document.createElement("foo")
  el.setAttribute("a", "1")
  el.setAttribute("b", "2")
  el.setAttribute("a", "3")
  el.setAttribute("c", "4")
  attributes_are(el, [["a", "3"],
                      ["b", "2"],
                      ["c", "4"]])
}, "setAttribute should not change the order of previously set attributes.")
test(function() {
  var el = document.createElement("baz")
  el.setAttributeNS("ab", "attr", "fail")
  el.setAttributeNS("kl", "attr", "pass")
  el.setAttribute("attr", "pass")
  attributes_are(el, [["attr", "pass", "ab"],
                      ["attr", "pass", "kl"]])
}, "setAttribute should set the first attribute with the given name")

// Miscellaneous tests
test(function() {
  var el = document.createElement("iframe")
  el.setAttribute("src", "file:///home")
  assert_equals(el.getAttribute("src"), "file:///home")
}, "setAttribute should not throw even if a load is not allowed")
</script>
